Common Uncouple Recouple Reducer
