--- src/modules/tests/test-data/cat-c/uncouple-recouple/uncouple-recouple.cat-ce.reducer.ts
@@ -1,47 +1,0 @@
-import { CatCEUniqueTypes } from '@dvsa/mes-test-schema/categories/CE';
-import * as uncoupleRecoupleActions from '../../common/uncouple-recouple/uncouple-recouple.actions';
-import { CompetencyOutcome } from '../../../../../shared/models/competency-outcome';
-
-export const initialState: CatCEUniqueTypes.UncoupleRecouple = {};
-
-export function uncoupleRecoupleCatCEReducer(
-  state = initialState,
-  action: uncoupleRecoupleActions.Types,
-): CatCEUniqueTypes.UncoupleRecouple {
-  switch (action.type) {
-    case uncoupleRecoupleActions.TOGGLE_UNCOUPLE_RECOUPLE:
-      return {
-        ...state,
-        selected: !state.selected,
-      };
-    case uncoupleRecoupleActions.UNCOUPLE_RECOUPLE_ADD_DRIVING_FAULT:
-      return {
-        ...state,
-        fault: CompetencyOutcome.DF,
-        selected: true,
-      };
-    case uncoupleRecoupleActions.UNCOUPLE_RECOUPLE_ADD_SERIOUS_FAULT:
-      return {
-        ...state,
-        fault: CompetencyOutcome.S,
-        selected: true,
-      };
-    case uncoupleRecoupleActions.UNCOUPLE_RECOUPLE_ADD_DANGEROUS_FAULT:
-      return {
-        ...state,
-        fault: CompetencyOutcome.D,
-        selected: true,
-      };
-    case uncoupleRecoupleActions.UNCOUPLE_RECOUPLE_REMOVE_FAULT:
-      return {
-        selected: state.selected,
-      };
-    case uncoupleRecoupleActions.ADD_UNCOUPLE_RECOUPLE_COMMENT:
-      return {
-        ...state,
-        faultComments: action.comment,
-      };
-    default:
-      return state;
-  }
-}
